restruct-dwh-layer

+++ profcomff_definitions/ODS/rating.py
@@ -19,14 +19,16 @@ class Lecturer(Base):
     subject: Mapped[str | None] = mapped_column(comment="Список предметов преподавателя")
     avatar_link: Mapped[str | None] = mapped_column(comment="Ссылка на аватар преподавателя")
     timetable_id: Mapped[int] = mapped_column(comment="Идертификатор в timetable-api")
-    mark_weighted: Mapped[int] = mapped_column(comment="Взвешенная оценка преподавателя", default=0, server_default="0")
-    mark_kindness_weighted: Mapped[int] = mapped_column(
+    mark_weighted: Mapped[float] = mapped_column(
+        comment="Взвешенная оценка преподавателя", default=0, server_default="0"
+    )
+    mark_kindness_weighted: Mapped[float] = mapped_column(
         comment="Взвешенная доброта преподавателя", default=0, server_default="0"
     )
-    mark_clarity_weighted: Mapped[int] = mapped_column(
+    mark_clarity_weighted: Mapped[float] = mapped_column(
         comment="Взверешенная понятность преподавателя", default=0, server_default="0"
     )
-    mark_freebie_weighted: Mapped[int] = mapped_column(
+    mark_freebie_weighted: Mapped[float] = mapped_column(
         comment="Взвешенная халявность преподавателя", default=0, server_default="0"
     )
     rank: Mapped[int] = mapped_column(comment="Место в рейтинге", default=0, server_default="0")
